classes were split but there are some bugs
