+++ RocketApp/app/src/main/java/com/example/rocketapp/view/Histogram.java
@@ -40,25 +40,24 @@ import java.util.Date;
 
 public class Histogram extends AppCompatActivity {
 
-    BarChart barChart; //Source: https://www.youtube.com/watch?v=pi1tq-bp7uA
-    LineChart lineChart; //Source: https://www.youtube.com/watch?v=yrbgN2UvKGQ&list=PLFh8wpMiEi89LcBupeftmAcgDKCeC24bJ&ab_channel=SarthiTechnology
-    Experiment experiment;
+    private BarChart barChart; //Source: https://www.youtube.com/watch?v=pi1tq-bp7uA
+    private Experiment experiment;
 
+    public BarChart getBarChart() {
+        return barChart;
+    }
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.graph_activity);
+    public void setBarChart(BarChart barChart) {
+        this.barChart = barChart;
+    }
 
-        barChart = (BarChart) findViewById(R.id.histogram);
-        lineChart = (LineChart) findViewById(R.id.time_plot);
+    public Histogram(Experiment experiment){
+        this.experiment = experiment;
+    }
+
+    public void createHistogramView(ArrayList<Trial> trials){
         ArrayList<BarEntry> barEntries = new ArrayList<BarEntry>();
         ArrayList<Trial> done = new ArrayList<Trial>();
-        experiment = ExperimentManager.getExperiment(getIntent().getSerializableExtra("id"));
-        ArrayList<Trial> trials = (ArrayList<Trial>) experiment.getFilteredTrials();
-        timePlot(trials);
-
-
         for (int i = 0; i < trials.size(); i++) {
             int freq = 0;
             int found = 0;
@@ -115,60 +114,5 @@ public class Histogram extends AppCompatActivity {
         barChart.setData(barData);
         barChart.getDescription().setText("Experiment Trials");
         barChart.animateY(2000);
-
-    }
-
-    private void timePlot(ArrayList<Trial> trials) {
-        XAxis xAxis = lineChart.getXAxis();
-        xAxis.setValueFormatter(new ValueFormatter() {
-            @Override
-            public String getFormattedValue(float value) {
-                Date date = new Date((long) value);
-                DateFormat df = new SimpleDateFormat("dd/MM");
-                return (df).format(date);// xVal is a string array
-            }
-        });
-        ArrayList<Entry> dataValue = new ArrayList<Entry>();
-        ArrayList<Entry> mockData = new ArrayList<Entry>();
-        float mean = 0;
-        ArrayList<String> done = new ArrayList<>();
-        for (int i = 0; i < trials.size(); i++) {
-            mean = experiment.getMean(trials.get(i).getTimestamp().toDate());
-            System.out.println("the mean is " + mean);//testing getMean(Date), tests confirm it works.
-            Entry entry = new Entry(trials.get(i).getTimestamp().toDate().getTime(), mean);
-            DateFormat df = new SimpleDateFormat("dd/MM");
-            String sdf = (df).format(trials.get(i).getTimestamp().toDate().getTime());
-            int found = 0;
-            for (int k = 0; k < done.size(); k++) {
-                if(sdf.equals(done.get(k))){
-                    found = found + 1;
-                }
-            }
-            if(found==0){
-                dataValue.add(entry);
-                done.add(sdf);
-            }
-            else{
-                for(int j = 0; j<dataValue.size(); j++){
-                    if((sdf.equals((df).format(dataValue.get(j).getX())))&&(entry.getX()>dataValue.get(j).getX())){
-                        dataValue.set(j, entry);
-                    }
-                }
-            }
-        }
-
-        Collections.sort(dataValue, new Comparator<Entry>() {
-            @Override
-            public int compare(Entry entry, Entry t1) {
-                return (int) (entry.getX()-t1.getX());
-            }
-        });
-                LineDataSet lineDataSet = new LineDataSet(dataValue, "Data set 1");
-        ArrayList<ILineDataSet> dataSets = new ArrayList<>();
-        dataSets.add(lineDataSet);
-        LineData data = new LineData(dataSets);
-        lineChart.setData(data);
-        System.out.println(lineChart.getData().getDataSets());
-
     }
 }
