--- RocketApp/app/src/main/java/com/example/rocketapp/view/Timeplot.java
@@ -1,4 +1,0 @@
-package com.example.rocketapp.view;
-
-public class Timeplot {
-}
